feat: create HFDatasetSelector component and use it in dataset selection page
